Add files via upload
"This project was designed to solve the challenge of fragmented reporting. While Amazon provides raw transaction reports, they lack a consolidated view of profitability and regional performance. I built this dashboard to transform thousands of rows of shipping and sales data into a strategic tool for inventory and profit optimization."